Renamed files, moved initiative to general node file. Replaced with stock initiative file
